Ticket #45: Addressed various build warnings
git-svn-id: http://svn.osgeo.org/fdo/trunk@4443 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/ArcSDE/Src/Provider/ArcSDEFeatureCommand.h
@@ -25,7 +25,7 @@
 #include "Fdo.h"
 
 #ifdef SDE_UNICODE
-typedef std::map <std::wstring, short > PropsSdeTypeMap;
+typedef std::map <std::wstring, long > PropsSdeTypeMap;
 #endif
 
 template <class FDO_COMMAND> class ArcSDEFeatureCommand : public ArcSDECommand<FDO_COMMAND>
